backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.123.0",
    "livekit-agents[sarvam,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "mem0ai==1.0.0",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
]
